Revert NZSL-170: Export prerelease videos
We have determined that private videos should not be included in the glossvideo
export in any circumstance, since they are used for purposes other than
for publishing.

To avoid accidental export, this commit removes the ability to filter the export
to include private videos, while retaining the test and code style improvements
made during development of the original task.