fix network graph for glosses
issue was that if there were glossrelations to other datasets, those were included in the edges,
but not in the nodes.
that lead to edges to point non-existing target nodes.
it might be desireable at some point to display cross lexicon relations, then the queries need to be changed again.